+++ strongbox/cryptography.ino
@@ -8,7 +8,7 @@
 // Why is "e" signed ? ¯\_(ツ)_/¯
 // It can't be negative but meh, there's a check in the logigram so let's do it like this !
 
-uint16_t modexp(uint16_t a, int16_t e, uint16_t n) {
+uint32_t modexp(uint16_t a, int16_t e, uint16_t n) {
   // r is signed in order to be able to return -1 in case the impossible is true
   int16_t r;
   if (e < 0) {
